Remove upstash library and unused scripts
