+++ apps/web/app/api/admin/links/[linkId]/ban/route.ts
@@ -1,8 +1,7 @@
 import { withAdmin } from "@/lib/auth";
 import { updateConfig } from "@/lib/edge-config";
 import { prisma } from "@/lib/prisma";
-import { redis } from "@/lib/redis";
-import { formatRedisLink } from "@/lib/upstash";
+import { formatRedisLink, redis } from "@/lib/redis";
 import {
   LEGAL_USER_ID,
   LEGAL_WORKSPACE_ID,
