+++ apps/web/app/api/links/[linkId]/transfer/route.ts
@@ -2,8 +2,7 @@ import { getClicks } from "@/lib/analytics/clicks";
 import { DubApiError } from "@/lib/api/errors";
 import { withWorkspace } from "@/lib/auth";
 import { prisma } from "@/lib/prisma";
-import { redis } from "@/lib/redis";
-import { formatRedisLink } from "@/lib/upstash";
+import { formatRedisLink, redis } from "@/lib/redis";
 import z from "@/lib/zod";
 import { trace } from "@opentelemetry/api";
 import { waitUntil } from "@vercel/functions";
