+++ apps/web/package.json
@@ -37,8 +37,6 @@
     "@splinetool/runtime": "^1.2.8",
     "@stripe/stripe-js": "^1.44.1",
     "@team-plain/typescript-sdk": "^3.9.0",
-    "@upstash/ratelimit": "^1.1.3",
-    "@upstash/redis": "^1.25.1",
     "@vercel/analytics": "^1.0.0",
     "@vercel/edge": "^0.3.1",
     "@vercel/edge-config": "^0.4.1",
